Pig Dice Game
This is a 2 player game where one player plays at a time while the other waits.
You can roll the dice and hold your score as per your wish or even choose to go for the next attempt.
Total score will be added to give the winner. Chance switches when the dice rolls to 1.